Very big change. IMPORTANT: in the middle of adding the "complete task" functionality.  Everything is working, mostly, but the last step is to reorder the task list in place correctly when the user completes a task.  Currently, the completed task just stays where it is until you refresh.  Everything still needs to be properly reordered in the RECEIVE_TOGGLE_COMPLETE reducer case.
Otherwise, the big change is bringing in a bunch of the code from the redux basic tutorial, code to do with task lists.  This replaces the existing task list with a more advanced one (VisibleTaskList), adds the visibility footer, and all the requisite actions/reducers.  The visibility code isn't working yet, even though the UI is there. The "toggle complete" ajax and bounding actions are all in place and working.

Note that I still haven't added any other code around the "isFetching" state, code that could disable parts of the UI during the various ajax requests (other than the first big get).  Probably need to eventually do something about that.