Adds partial support for "completing/reviving" tasks.  All of this code is likely to change or be thrown out altogether, since I'm now going to replace the existing list with a sortable component, but I didn't want to lose this code, so I'm just going to check it in for now.
The most significant part of this change is a mass of logic in the reducer for the RECEIVE_TOGGLE_COMPLETE case -- this code takes a recently completed task and moves it to the top of the list of completed tasks.  Again, this is really dead code at this point, but just to explain a little.