Things are still a mess, but not quite as bad.  The new react-sortable-hoc is working out much better -- at least, I see a way through the stateful/stateless component disconnect.  The new component provides an obvious way to respond to the drag n' drop reordering, without breaking the entire use of redux.
Still, it's feeling a bit awkward.  The current strategy will be to update the server in response to list change of order.  If that's successful, update the client state to reflect the server state, and continue.  This is a little nuts:  we're replicating the logic of the server to the client.  This is especially a problem from a maintenance pov.  I don't like it, but the alternative is to do a complete fetch of all tasks after any order change.  This causes a noticeable flicker update in the UI; might make more sense to have overlay a spinner?

Anyhow, lots to do to fix this. The server update is working fine, but the reducer code to match the server still needs to be written.  Hmmmm.